refactor: extract aegis http feature into ferriskey-api-aegis
