+++ libs/ferriskey-api-aegis/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "ferriskey-api-aegis"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-api-core = { path = "../ferriskey-api-core" }
+ferriskey-core = { path = "../../core" }
+axum = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+utoipa = { workspace = true }
+uuid = { workspace = true }
+validator = { workspace = true }
