+++ libs/ferriskey-api-aegis/src/handlers.rs
