+++ libs/ferriskey-api-aegis/src/handlers/assign_optional_scope.rs
@@ -1,11 +1,9 @@
-use crate::application::http::server::{
-    api_entities::{api_error::ApiError, response::Response},
-    app_state::AppState,
-};
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::{api_error::ApiError, response::Response};
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::aegis::entities::ClientScopeMapping;
 use ferriskey_core::domain::aegis::ports::ScopeMappingService;
 use ferriskey_core::domain::aegis::value_objects::AssignClientScopeInput;
@@ -14,9 +12,9 @@ use uuid::Uuid;
 
 #[utoipa::path(
     put,
-    path = "/clients/{client_id}/default-client-scopes/{scope_id}",
-    summary = "Assign a default client scope to a client",
-    description = "Assigns a client scope as a default scope to the specified client. Default scopes are always included in tokens.",
+    path = "/clients/{client_id}/optional-client-scopes/{scope_id}",
+    summary = "Assign an optional client scope to a client",
+    description = "Assigns a client scope as an optional scope to the specified client. Optional scopes are included only when explicitly requested.",
     params(
         ("realm_name" = String, Path, description = "Realm name"),
         ("client_id" = Uuid, Path, description = "Client ID"),
@@ -24,10 +22,10 @@ use uuid::Uuid;
     ),
     tag = "client-scope",
     responses(
-        (status = 200, body = ClientScopeMapping, description = "Default scope assigned successfully"),
+        (status = 200, body = ClientScopeMapping, description = "Optional scope assigned successfully"),
     ),
 )]
-pub async fn assign_default_scope(
+pub async fn assign_optional_scope(
     Path((realm_name, client_id, scope_id)): Path<(String, Uuid, Uuid)>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
@@ -40,8 +38,8 @@ pub async fn assign_default_scope(
                 realm_name,
                 client_id,
                 scope_id,
-                is_default: true,
-                is_optional: false,
+                is_default: false,
+                is_optional: true,
             },
         )
         .await?;
