+++ libs/ferriskey-api-aegis/src/handlers/create_client_scope.rs
@@ -1,17 +1,13 @@
-use crate::application::http::{
-    aegis::validators::CreateClientScopeValidator,
-    server::{
-        api_entities::{
-            api_error::{ApiError, ValidateJson},
-            response::Response,
-        },
-        app_state::AppState,
-    },
-};
+use crate::validators::CreateClientScopeValidator;
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ValidateJson},
+    response::Response,
+};
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::aegis::entities::ClientScope;
 use ferriskey_core::domain::aegis::ports::ClientScopeService;
 use ferriskey_core::domain::aegis::value_objects::CreateClientScopeInput;
