+++ libs/ferriskey-api-aegis/src/handlers/get_client_client_scopes.rs
@@ -1,14 +1,12 @@
-use crate::application::http::server::{
-    api_entities::{
-        api_error::{ApiError, ApiErrorResponse},
-        response::Response,
-    },
-    app_state::AppState,
-};
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ApiErrorResponse},
+    response::Response,
+};
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::aegis::entities::ClientScope;
 use ferriskey_core::domain::aegis::ports::ScopeMappingService;
 use ferriskey_core::domain::aegis::value_objects::GetClientClientScopesInput;
